+++ NEWS.md
@@ -4,6 +4,7 @@
 * labelling and popups for routes
 * supplying multiple travel modes using recursion
 * formatting options for input times and distances
+* wrap line-drawing functions
 
 # TargomoR 0.0.0.9005